Support edit username in username screen
